LPD-31609 Group page editor tests better
