+++ modules/test/playwright/tests/layout-content-page-editor-web/collectionDisplayFragment.spec.ts
