--- modules/test/playwright/tests/layout-content-page-editor-web/htmlFragment.spec.ts
@@ -1,60 +1,0 @@
-/**
- * SPDX-FileCopyrightText: (c) 2000 Liferay, Inc. https://liferay.com
- * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
- */
-
-import {expect, mergeTests} from '@playwright/test';
-
-import {apiHelpersTest} from '../../fixtures/apiHelpersTest';
-import {featureFlagsTest} from '../../fixtures/featureFlagsTest';
-import {isolatedSiteTest} from '../../fixtures/isolatedSiteTest';
-import {loginTest} from '../../fixtures/loginTest';
-import {pageEditorPagesTest} from '../../fixtures/pageEditorPagesTest';
-import getRandomString from '../../utils/getRandomString';
-import getFragmentDefinition from './utils/getFragmentDefinition';
-import getPageDefinition from './utils/getPageDefinition';
-
-const test = mergeTests(
-	apiHelpersTest,
-	featureFlagsTest({
-		'LPS-178052': true,
-	}),
-	isolatedSiteTest,
-	loginTest(),
-	pageEditorPagesTest
-);
-
-test('Can edit html editable', async ({
-	apiHelpers,
-	page,
-	pageEditorPage,
-	site,
-}) => {
-
-	// Create page with a heading fragment and go to edit mode
-
-	const headingId = getRandomString();
-
-	const headingFragment = getFragmentDefinition({
-		id: headingId,
-		key: 'BASIC_COMPONENT-html',
-	});
-
-	const layout = await apiHelpers.headlessDelivery.createSitePage({
-		pageDefinition: getPageDefinition([headingFragment]),
-		siteId: site.id,
-		title: getRandomString(),
-	});
-
-	await pageEditorPage.goto(layout, site.friendlyUrlPath);
-
-	// Check html editable can be edited
-
-	await pageEditorPage.editHTMLEditable(
-		headingId,
-		'element-html',
-		'<div class="text-success"><h1>test html</h1></div>'
-	);
-
-	await expect(page.getByText('test html')).toBeAttached();
-});
