--- modules/test/playwright/tests/layout-content-page-editor-web/menuDisplayWidget.spec.ts
@@ -1,69 +1,0 @@
-/**
- * SPDX-FileCopyrightText: (c) 2000 Liferay, Inc. https://liferay.com
- * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
- */
-
-import {expect, mergeTests} from '@playwright/test';
-
-import {apiHelpersTest} from '../../fixtures/apiHelpersTest';
-import {featureFlagsTest} from '../../fixtures/featureFlagsTest';
-import {isolatedSiteTest} from '../../fixtures/isolatedSiteTest';
-import {loginTest} from '../../fixtures/loginTest';
-import getRandomString from '../../utils/getRandomString';
-import getPageDefinition from './utils/getPageDefinition';
-import getWidgetDefinition from './utils/getWidgetDefinition';
-
-const test = mergeTests(
-	apiHelpersTest,
-	featureFlagsTest({
-		'LPS-178052': true,
-	}),
-	isolatedSiteTest,
-	loginTest()
-);
-
-test('Checks that the Display Menu items have a role link with display style Bar With Links', async ({
-	apiHelpers,
-	page,
-	site,
-}) => {
-	const widgetDefinition = getWidgetDefinition({
-		id: getRandomString(),
-		widgetConfig: {
-			displayStyle: 'ddmTemplate_NAVBAR-LINKS-FTL',
-		},
-		widgetName:
-			'com_liferay_site_navigation_menu_web_portlet_SiteNavigationMenuPortlet',
-	});
-
-	// Create three pages, one of them with Menu Display widget
-
-	await apiHelpers.headlessDelivery.createSitePage({
-		siteId: site.id,
-		title: 'First page',
-	});
-
-	const secondLayout = await apiHelpers.headlessDelivery.createSitePage({
-		siteId: site.id,
-		title: 'Second page',
-	});
-
-	const thirdLayout = await apiHelpers.headlessDelivery.createSitePage({
-		pageDefinition: getPageDefinition([widgetDefinition]),
-		parentSitePage: {friendlyUrlPath: secondLayout.friendlyUrlPath},
-		siteId: site.id,
-		title: 'Third page',
-	});
-
-	await page.goto(
-		`/web${site.friendlyUrlPath}${thirdLayout.friendlyUrlPath}`
-	);
-
-	// Check that the Display Menu items have a role link
-
-	await page.getByRole('link', {name: 'Second page'}).hover();
-
-	await expect(page.getByRole('link', {name: 'First page'})).toBeVisible();
-	await expect(page.getByRole('link', {name: 'Second page'})).toBeVisible();
-	await expect(page.getByRole('link', {name: 'Third page'})).toBeVisible();
-});
