--- modules/test/playwright/tests/layout-content-page-editor-web/rules.spec.ts
@@ -1,65 +1,0 @@
-/**
- * SPDX-FileCopyrightText: (c) 2000 Liferay, Inc. https://liferay.com
- * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
- */
-
-import {mergeTests} from '@playwright/test';
-
-import {apiHelpersTest} from '../../fixtures/apiHelpersTest';
-import {featureFlagsTest} from '../../fixtures/featureFlagsTest';
-import {isolatedSiteTest} from '../../fixtures/isolatedSiteTest';
-import {loginTest} from '../../fixtures/loginTest';
-import {pageEditorPagesTest} from '../../fixtures/pageEditorPagesTest';
-import {checkAccessibility} from '../../utils/checkAccessibility';
-import getRandomString from '../../utils/getRandomString';
-import getFragmentDefinition from './utils/getFragmentDefinition';
-import getPageDefinition from './utils/getPageDefinition';
-
-const test = mergeTests(
-	apiHelpersTest,
-	featureFlagsTest({
-		'LPS-169837': true,
-		'LPS-178052': true,
-	}),
-	isolatedSiteTest,
-	loginTest(),
-	pageEditorPagesTest
-);
-
-test('Checks the accessibility of the rule modal by filling out a condition and an action', async ({
-	apiHelpers,
-	page,
-	pageEditorPage,
-	site,
-}) => {
-
-	// Create content page with a Heading fragment and go to edit mode
-
-	const layout = await apiHelpers.headlessDelivery.createSitePage({
-		pageDefinition: getPageDefinition([
-			getFragmentDefinition({
-				id: getRandomString(),
-				key: 'BASIC_COMPONENT-heading',
-			}),
-		]),
-		siteId: site.id,
-		title: getRandomString(),
-	});
-
-	await pageEditorPage.goto(layout, site.friendlyUrlPath);
-
-	// Add rule and check accessibility of modal
-
-	await pageEditorPage.goToSidebarTab('Page Rules');
-
-	await page.getByRole('button', {name: 'New Rule'}).click();
-
-	await pageEditorPage.addRuleCondition();
-
-	await pageEditorPage.addRuleAction();
-
-	await checkAccessibility({
-		page,
-		selectors: ['.modal-body'],
-	});
-});
