+++ modules/test/playwright/tests/layout-content-page-editor-web/sidebar.spec.ts
@@ -8,11 +8,14 @@ import {expect, mergeTests} from '@playwright/test';
 import {apiHelpersTest} from '../../fixtures/apiHelpersTest';
 import {applicationsMenuPageTest} from '../../fixtures/applicationsMenuPageTest';
 import {featureFlagsTest} from '../../fixtures/featureFlagsTest';
+import {fragmentsPagesTest} from '../../fixtures/fragmentPagesTest';
 import {isolatedSiteTest} from '../../fixtures/isolatedSiteTest';
 import {loginTest} from '../../fixtures/loginTest';
 import {pageEditorPagesTest} from '../../fixtures/pageEditorPagesTest';
 import {checkAccessibility} from '../../utils/checkAccessibility';
 import getRandomString from '../../utils/getRandomString';
+import getFragmentDefinition from './utils/getFragmentDefinition';
+import getPageDefinition from './utils/getPageDefinition';
 
 const test = mergeTests(
 	apiHelpersTest,
@@ -21,6 +24,7 @@ const test = mergeTests(
 		'LPS-169837': true,
 		'LPS-178052': true,
 	}),
+	fragmentsPagesTest,
 	isolatedSiteTest,
 	loginTest(),
 	pageEditorPagesTest
@@ -175,5 +179,156 @@ test('Checks sidebar accessibility', async ({
 	await checkAccessibility({
 		page,
 		selectors: ['.page-editor__sidebar'],
+	});
+});
+
+test.describe('Fragments Panel', () => {
+	test('Only published fragments are shown in the Fragments Sidebar', async ({
+		apiHelpers,
+		fragmentEditorPage,
+		fragmentsPage,
+		page,
+		pageEditorPage,
+		site,
+	}) => {
+
+		// Create new fragment set
+
+		await fragmentsPage.goto(site.friendlyUrlPath);
+
+		const setName = getRandomString();
+		await fragmentsPage.createFragmentSet(setName);
+
+		// Create unpublished fragment inside it
+
+		await fragmentsPage.goto(site.friendlyUrlPath);
+
+		const unpublishedFragmentName = getRandomString();
+		await fragmentsPage.createFragment(setName, unpublishedFragmentName);
+
+		// Create published fragment inside it
+
+		await fragmentsPage.goto(site.friendlyUrlPath);
+
+		const publishedFragmentName = getRandomString();
+		await fragmentsPage.createFragment(setName, publishedFragmentName);
+		await fragmentEditorPage.publish();
+
+		// Create content page and go to edit mode
+
+		const layout = await apiHelpers.headlessDelivery.createSitePage({
+			pageDefinition: getPageDefinition(),
+			siteId: site.id,
+			title: getRandomString(),
+		});
+
+		await pageEditorPage.goto(layout, site.friendlyUrlPath);
+
+		// Check only published fragment is displayed
+
+		await pageEditorPage.goToSidebarTab('Fragments and Widgets');
+
+		await page
+			.getByRole('menuitem', {
+				exact: true,
+				name: setName,
+			})
+			.click();
+
+		await expect(page.getByText(publishedFragmentName)).toBeVisible();
+
+		await expect(page.getByText(unpublishedFragmentName)).not.toBeVisible();
+	});
+});
+
+test.describe('Page Contents Panel', () => {
+	test('Allows editing inline text from Page Content Panel', async ({
+		apiHelpers,
+		page,
+		pageEditorPage,
+		site,
+	}) => {
+
+		// Create a page with a Heading fragment
+
+		const headingId = getRandomString();
+		const headingDefinition = getFragmentDefinition({
+			id: headingId,
+			key: 'BASIC_COMPONENT-heading',
+		});
+
+		const layout = await apiHelpers.headlessDelivery.createSitePage({
+			pageDefinition: getPageDefinition([headingDefinition]),
+			siteId: site.id,
+			title: getRandomString(),
+		});
+
+		// Go to edit mode of page
+
+		await pageEditorPage.goto(layout, site.friendlyUrlPath);
+
+		// Go to Page Contents panel and edit inline text
+
+		await pageEditorPage.goToSidebarTab('Page Content');
+
+		await page.getByLabel('Edit Text Heading Example').click();
+
+		const editable = pageEditorPage.getEditable(headingId, 'element-text');
+
+		await editable.locator('.cke_editable_inline').waitFor();
+
+		// Clear current content and fill with new one
+
+		await page.keyboard.press('Control+KeyA');
+		await page.keyboard.press('Backspace');
+
+		await page.keyboard.type('New Content');
+		await page.locator('body').click();
+
+		await pageEditorPage.waitForChangesSaved();
+
+		await expect(
+			page.locator('.page-editor__page-contents__page-content')
+		).toContainText('New Content');
+	});
+});
+
+test.describe('Rules Panel', () => {
+	test('Checks the accessibility of the rule modal by filling out a condition and an action', async ({
+		apiHelpers,
+		page,
+		pageEditorPage,
+		site,
+	}) => {
+
+		// Create content page with a Heading fragment and go to edit mode
+
+		const layout = await apiHelpers.headlessDelivery.createSitePage({
+			pageDefinition: getPageDefinition([
+				getFragmentDefinition({
+					id: getRandomString(),
+					key: 'BASIC_COMPONENT-heading',
+				}),
+			]),
+			siteId: site.id,
+			title: getRandomString(),
+		});
+
+		await pageEditorPage.goto(layout, site.friendlyUrlPath);
+
+		// Add rule and check accessibility of modal
+
+		await pageEditorPage.goToSidebarTab('Page Rules');
+
+		await page.getByRole('button', {name: 'New Rule'}).click();
+
+		await pageEditorPage.addRuleCondition();
+
+		await pageEditorPage.addRuleAction();
+
+		await checkAccessibility({
+			page,
+			selectors: ['.modal-body'],
+		});
 	});
 });
